Move constant reconstruction to its own module
